requirements.txt
django~=3.0.2
djangorestframework~=3.13.1
pika~=1.2.0
python-decouple